requirements.txt
numpy
opencv-contrib-python==4.8.1.78
openpyxl
pandas
pillow
pyttsx3